feat(script): subject type
