+++ script/config.py
@@ -1,0 +1,3 @@
+CAP_BASE_URL: str = "https://cap.rcpet.edu.tw/exam/"
+CAP_START_YEAR: int = 103
+CAP_END_YEAR: int = 111
