+++ script/models/subject.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class CAPSubject(Enum):
+    CHINESE = "Chinese"
+    ENGLISH = "English"
+    MATH = "Math"
+    NATURE = "Nature"
+    SOCIETY = "Society"
