+++ script/pyproject.toml
@@ -3,7 +3,7 @@ name = "cap-countdown-script"
 version = "0.1.0"
 description = "A script tool to fetch and tidy up the data from the CAP website."
 authors = ["SiongSng <rrt467778@gmail.com>"]
-packages = [{include = "tools"}]
+packages = [{include = "tools"}, {include = "models"}]
 license = "GPL-3.0-only"
 repository="https://github.com/SiongSng/cap-countdown"
